refactor(layout): move monitor/+recover/ to training/mlops/ (T124.4.4)
